Large test whooooooooooo

+++ lexer/src/types.rs
@@ -42,7 +42,7 @@ impl Range {
 pub enum TokenType {
     Ident(String),
     String(String),
-    Number(f32),
+    Number(f64),
     // Comments probably don't need to contain what's in the comment, but I'll leave this for now
     Comment(String),
